add sidebar content fetching and fix name
