+++ index.ts
@@ -3,11 +3,19 @@ import { StdioServerTransport } from "@modelcontextprotocol/sdk/server/stdio.js"
 import "dotenv/config";
 import { getGuide, testAgentResponse } from "./tools.js";
 import { getGuideParams, testAgentResponseParams } from "./params.js";
+import { fetchAndUpdateSidebar } from "./sidebar.js";
 
+// Initialize the server
 const server = new McpServer({
   name: "docs-mcp",
   version: "1.0.0",
 });
+
+// Fetch sidebar before starting the server
+await fetchAndUpdateSidebar();
+
+// Set up a periodic update (e.g., every hour)
+setInterval(fetchAndUpdateSidebar, 60 * 60 * 1000);
 
 server.tool(
   "BuildOnBase",
@@ -22,5 +30,6 @@ server.tool(
   testAgentResponseParams.shape,
   testAgentResponse
 );
+
 const transport = new StdioServerTransport();
 server.connect(transport);
